Altitude PID tuning, CommLink is now static, motor test is pulsed
CommLink is now a static class (for size, speed).  MotorTest is now
timed instead of requiring on / off commands (safer).  Altitude PIDs are
tuned reasonably well now.  Still getting weird behavior with integral.
Moved logging to its own cog, but commented it out.  Cycle times are
converted to /64 units to reduce transmit size, memory cost.

+++ Elev8-Groundstation/Elev8Data.cs
@@ -102,15 +102,15 @@ namespace Elev8
 
 	public class DebugValues
 	{
-		public int MinCycles, MaxCycles, AvgCycles;
+		public short MinCycles, MaxCycles, AvgCycles;
 		public int Counter;
 		public float DebugFloat;
 
 		public void ReadFrom( Packet p )
 		{
-			MinCycles = p.GetInt();
-			MaxCycles = p.GetInt();
-			AvgCycles = p.GetInt();
+			MinCycles = p.GetShort();
+			MaxCycles = p.GetShort();
+			AvgCycles = p.GetShort();
 			Counter = p.GetInt();		// basically a sequence value
 			DebugFloat = p.GetFloat();
 		}
